# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LearnAI - AI-Powered Adaptive Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List all uploaded notes
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Upload a note / study material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/upload:
    post:
      operationId: uploadNote
      tags: [notes]
      summary: Upload a PDF, image, or text file as study material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUploadInput"
      responses:
        "201":
          description: Created note from uploaded file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a specific note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ─── COURSES ─────────────────────────────────────────────────────────────────
...
  /exam/mock:
    get:
      operationId: getExamMock
      tags: [tasks]
      summary: Generate a timed exam simulation from weak concepts
      parameters:
        - name: count
          in: query
          schema:
            type: integer
            default: 10
        - name: courseId
          in: query
          schema:
            type: integer
        - name: durationMinutes
          in: query
          schema:
            type: integer
            default: 25
      responses:
        "200":
          description: Exam mock session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamMock"
        "404":
          description: No courses available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ─── OPENAI ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── NOTE ──────────────────────────────────────────────────────────────────
    Note:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        subject:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        mimeType:
          type: ["string", "null"]
        sourceType:
          type: string
          enum: [text, pdf, image]
        wordCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - wordCount
        - createdAt
        - updatedAt
    NoteInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 10
        subject:
          type: string
      required:
        - title
        - content
    NoteUploadInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        subject:
          type: string
        fileName:
          type: string
          minLength: 1
        mimeType:
          type: string
        fileBase64:
          type: string
          minLength: 1
      required:
        - title
        - fileName
        - fileBase64

    # ─── COURSE ────────────────────────────────────────────────────────────────
...
    ExamMockQuestion:
      type: object
      properties:
        index:
          type: integer
        stepId:
          type: integer
        courseId:
          type: integer
        courseTitle:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        question:
          type: string
        type:
          type: string
        choices:
          type: ["array", "null"]
          items:
            type: string
      required:
        - index
        - stepId
        - courseId
        - question
        - type
    ExamMock:
      type: object
      properties:
        durationMinutes:
          type: integer
        questionCount:
          type: integer
        startedAt:
          type: string
          format: date-time
        questions:
          type: array
          items:
            $ref: "#/components/schemas/ExamMockQuestion"
      required:
        - durationMinutes
        - questionCount
        - startedAt
        - questions